+++ test/mysql/columns.test.ts
@@ -1,6 +1,6 @@
 import {suite} from '@alinea/suite'
 import * as mysql from '@/mysql/columns.ts'
-import {columnSql, mapFrom, mapTo} from '../../test/TestUtils.ts'
+import {columnSql, mapFrom, mapTo} from '../TestUtils.ts'
 
 const test = suite(import.meta)
