+++ test/sqlite/columns.test.ts
@@ -1,6 +1,6 @@
 import {suite} from '@alinea/suite'
 import * as sqlite from '@/sqlite/columns.ts'
-import {columnSql, mapFrom, mapTo} from '../../test/TestUtils.ts'
+import {columnSql, mapFrom, mapTo} from '../TestUtils.ts'
 
 const test = suite(import.meta)
